move to syncronous execution of combos
