+++ packages/vue-keystrokes/src/tests/use-key-combo.spec.ts
@@ -50,7 +50,6 @@ describe('useKeyCombo(keyCombo) -> isPressed', () => {
 
     keystrokes.press({ key: 'a' })
     keystrokes.press({ key: 'b' })
-
     await wait()
 
     expect(w.get('div').text()).toEqual('isPressed')
@@ -65,8 +64,6 @@ describe('useKeyCombo(keyCombo) -> isPressed', () => {
 
     keystrokes.press({ key: 'a' })
     keystrokes.press({ key: 'b' })
-    await wait()
-
     keystrokes.release({ key: 'a' })
     keystrokes.release({ key: 'b' })
     await wait()
